pyproject.toml
[project]
name = "photon-platform-clerk"
dynamic = ["version"]
authors = [
  { name="phi ARCHITECT", email="github@phiarchitect.com" },
]
description = "the ultimate assistant"
readme = "README.rst"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "textual",
    "rich",
    "jinja2",
    "click",
    "requests",
    "beautifulsoup4",
    "openai",
    "python-slugify",
]

[project.scripts]
clerk = "photon_platform.clerk.__main__:run"
rollup = "photon_platform.clerk.rollup:rollup"
status_diff = "photon_platform.clerk.status_diff:status_diff"

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["photon_platform"]
namespaces = true

[tool.setuptools.dynamic]
version = {attr = "photon_platform.clerk.__version__"}

[project.urls]
"Website" = "https://photon-platform.github.io/clerk"
"Repository" = "https://github.com/photon-platform/clerk"
"Issues" = "https://github.com/photon-platform/clerk/issues"
